+++ src/rnn/layouts/network.rs
@@ -6,8 +6,9 @@ use std::sync::atomic::{AtomicUsize, Ordering};
 use std::sync::Arc;
 use std::time::Duration;
 
+use chrono::Utc;
 use regex::Regex;
-use tokio::sync::broadcast::{self, Receiver, Sender};
+use tokio::sync::broadcast::{self, Receiver};
 use tokio::sync::{mpsc, RwLock};
 use tokio::time;
 use tokio_util::sync::CancellationToken;
@@ -19,9 +20,11 @@ use crate::rnn::common::network_cfg::NeuronCfg;
 use crate::rnn::common::rnn_error::RnnError;
 use crate::rnn::common::signal::{Signal, Weight};
 use crate::rnn::common::spec_type::SpecificationType;
-use crate::rnn::common::status::Status;
+use crate::rnn::common::status::{PortInfo, Status};
 use crate::rnn::common::utils::gen_id_by_spec_type;
-use crate::rnn::neural::neuron::{self, Neuron};
+use crate::rnn::neural::neuron::Neuron;
+
+use super::signal_handler::SignalHandler;
 
 static ID_COUNTER: AtomicUsize = AtomicUsize::new(0_usize);
 static CHANNEL_CAPACITY: usize = 5;
@@ -44,27 +47,60 @@ struct Modes {
     monitoring_mode: MonitoringMode,
 }
 
+/// Network spreads command via command channel to all neurons.
 #[derive(Debug)]
 struct CommandsCh {
     sender: Arc<broadcast::Sender<NeuronCommand>>,
 }
 
+/// Neurons and ports are sending status information to Network when MonitoringMode is enabled
 #[derive(Debug)]
 struct MonitoringCh {
     sender: Arc<mpsc::Sender<Status>>,
     store: Arc<RwLock<Vec<Status>>>,
 }
 
+/// The port proprties
+#[derive(Debug)]
+struct PortCore {
+    /// Port id
+    id: String,
+
+    /// Signal hits counter
+    signal_hits: u64,
+
+    /// Sender or receiver handler dependent from port kind (Input or Output)
+    signal_handler: SignalHandler,
+}
+
 /// Network is a high level container to other containers (neurons)
 #[derive(Debug)]
 pub struct Network {
+    /// The network id
     id: String,
+
+    /// Inner neurons
     neurons: RwLock<BTreeMap<String, Arc<Neuron>>>,
+
+    /// Network's modes set
     modes: Arc<RwLock<Modes>>,
-    input_interface: Arc<RwLock<BTreeMap<usize, Arc<RwLock<Sender<u8>>>>>>,
-    output_interface: Arc<RwLock<BTreeMap<usize, Arc<RwLock<Receiver<u8>>>>>>,
+
+    /// Network's input ports
+    input_interface: Arc<RwLock<BTreeMap<usize, Arc<RwLock<PortCore>>>>>,
+
+    /// Network's output ports
+    output_interface: Arc<RwLock<BTreeMap<usize, Arc<RwLock<PortCore>>>>>,
+
+    /// The command channel stuff. Network sends command to all inner neurons
+    /// through broadcast channel.
     commands_ch: CommandsCh,
+
+    /// The monitoring channel stuff. Neurons are sending staus messages to
+    /// the network when enabled monitoring mode
     monitoring_ch: MonitoringCh,
+
+    /// all receivers are handling incoming signals and messages in separated tasks.
+    /// These tasks are tracked by receivers_tracker.
     receivers_tracker: TaskTracker,
     cancel_token: CancellationToken,
 }
@@ -88,7 +124,6 @@ impl Network {
             input_interface: Arc::new(RwLock::new(BTreeMap::new())),
             output_interface: Arc::new(RwLock::new(BTreeMap::new())),
             commands_ch: CommandsCh {
-                // receiver: Arc::new(commands_receiver), // TODO: Try not keeping this because receivers created as subscriptions
                 sender: Arc::new(commands_sender),
             },
             monitoring_ch: MonitoringCh {
@@ -186,10 +221,7 @@ impl Network {
             Entry::Vacant(entry) => Ok(Arc::clone(
                 entry.insert(Neuron::build(Arc::clone(&network), neuron_config).await),
             )),
-            Entry::Occupied(_) => Err(Box::new(RnnError::IdBusy(format!(
-                "Id {} already used",
-                new_id
-            )))),
+            Entry::Occupied(_) => Err(Box::new(RnnError::NeuronAlreadyExists(new_id))),
         }
     }
 
@@ -216,7 +248,7 @@ impl Network {
         let mut w_neurons = self.neurons.write().await;
         match w_neurons.remove(id) {
             Some(_) => Ok(()),
-            None => Err(Box::new(RnnError::IdNotFound)),
+            None => Err(Box::new(RnnError::NeuronNotFound(id.to_string()))),
         }
     }
 
@@ -231,36 +263,52 @@ impl Network {
         dst_port: usize,
     ) -> Result<(), Box<dyn Error>> {
         let src_neuron = self.get_neuron(src_id).await;
-        let dst_neuron = self.get_neuron(dst_id).await;
-        if src_neuron.is_some() && dst_neuron.is_some() {
-            let src_neuron = src_neuron.unwrap();
-            let dst_neuron = dst_neuron.unwrap();
-            src_neuron.link_to(dst_neuron, dst_port).await
-        } else {
-            Err(Box::new(RnnError::IdNotFound))
+        if src_neuron.is_none() {
+            return Err(Box::new(RnnError::NeuronNotFound(src_id.to_string())));
         }
+
+        let dst_neuron = self.get_neuron(dst_id).await;
+        if dst_neuron.is_none() {
+            return Err(Box::new(RnnError::NeuronNotFound(dst_id.to_string())));
+        }
+
+        let src_neuron = src_neuron.unwrap();
+        let dst_neuron = dst_neuron.unwrap();
+        src_neuron.link_to(dst_neuron, dst_port).await
     }
 
     pub async fn len(&self) -> usize {
         self.neurons.read().await.len()
     }
 
-    // pub fn send_statistics(&self, stc_id: &str)
-
     /// Send signal to port connected to synapse
     pub async fn input(&self, signal: Signal, port: usize) -> Result<usize, Box<dyn Error>> {
-        if let Some(synapse) = self.input_interface.read().await.get(&port) {
-            {
-                if self.get_monitoring_mode().await == MonitoringMode::Monitoring {
-                    // self.send_statistics()
+        if let Some(port_core) = self.input_interface.write().await.get_mut(&port) {
+            let mut w_port_core = port_core.write().await;
+            w_port_core.signal_hits += 1;
+            if let SignalHandler::Input(synapse) = &w_port_core.signal_handler {
+                let w_synapse = synapse.read().await;
+                let result = w_synapse
+                    .send(signal)
+                    .map_err(|error| Box::new(error) as Box<dyn Error>);
+
+                if result.is_ok() && self.get_monitoring_mode().await == MonitoringMode::Monitoring
+                {
+                    Self::send_port_status(
+                        self.monitoring_ch.store.clone(),
+                        w_port_core.id.as_str(),
+                        w_port_core.signal_hits,
+                        signal,
+                    )
+                    .await;
                 }
+
+                result
+            } else {
+                Err(Box::new(RnnError::IncorrectPortType))
             }
-            let r_synapse = synapse.read().await;
-            r_synapse
-                .send(signal)
-                .map_err(|err| Box::new(err) as Box<dyn Error>)
         } else {
-            Err(Box::new(RnnError::IdNotFound))
+            Err(Box::new(RnnError::PortNotFound(port)))
         }
     }
 
@@ -283,16 +331,17 @@ impl Network {
             let mut w_input_interface = self.input_interface.write().await;
             match w_input_interface.entry(network_port) {
                 Entry::Vacant(entry) => {
-                    entry.insert(Arc::new(RwLock::new(tx)));
+                    entry.insert(Arc::new(RwLock::new(PortCore {
+                        id: src_id.clone(),
+                        signal_hits: 0,
+                        signal_handler: SignalHandler::Input(Arc::new(RwLock::new(tx))),
+                    })));
                     Ok(())
                 }
-                Entry::Occupied(_) => Err(Box::new(RnnError::IdBusy(format!(
-                    "network port {} already used",
-                    network_port
-                )))),
+                Entry::Occupied(_) => Err(Box::new(RnnError::PortBusy(src_id))),
             }
         } else {
-            Err(Box::new(RnnError::IdNotFound) as Box<dyn Error>)
+            Err(Box::new(RnnError::NeuronNotFound(neuron_id.to_string())) as Box<dyn Error>)
         }
     }
 
@@ -303,14 +352,14 @@ impl Network {
         snapshot
     }
 
-    pub async fn get_current_neuron_statistics(
+    pub async fn get_current_neuron_status(
         &self,
         neuron_id: &str,
     ) -> Result<Status, Box<dyn Error>> {
         if let Some(target) = self.get_neuron(neuron_id).await {
             Ok(Neuron::prepare_status(&target.get_id(), &target.get_core()).await)
         } else {
-            Err(Box::new(RnnError::IdNotFound))
+            Err(Box::new(RnnError::NeuronNotFound(String::from(neuron_id))))
         }
     }
 
@@ -321,7 +370,7 @@ impl Network {
                 entity.remove();
                 Ok(())
             }
-            Entry::Vacant(_) => Err(Box::new(RnnError::AlreadyFree)),
+            Entry::Vacant(_) => Err(Box::new(RnnError::PortAlreadyFree)),
         }
     }
 
@@ -330,57 +379,76 @@ impl Network {
         network_port: usize,
         neuron_id: &str,
     ) -> Result<(), Box<dyn Error>> {
+        let port_id = format!("{}O{}", self.get_id(), network_port);
         let mut w_output_interface = self.output_interface.write().await;
         if let Some(neuron) = self.get_neuron(neuron_id).await {
             // connect axon to output port if port is free else return error
             match w_output_interface.entry(network_port.clone()) {
-                Entry::Occupied(_) => Err(Box::new(RnnError::IdBusy(format!(
-                    "Port {} already used",
-                    network_port
-                )))),
-                Entry::Vacant(entity) => {
-                    entity.insert(neuron.provide_output().await);
+                Entry::Occupied(_) => Err(Box::new(RnnError::PortBusy(port_id.clone()))),
+                Entry::Vacant(entry) => {
+                    let receiver = neuron.provide_output().await;
+                    let port_core = Arc::new(RwLock::new(PortCore {
+                        id: port_id.clone(),
+                        signal_hits: 0,
+                        signal_handler: SignalHandler::Output(receiver.clone()),
+                    }));
+                    entry.insert(port_core.clone());
+                    let monitoring_store_cloned = self.monitoring_ch.store.clone();
+
+                    self.receivers_tracker.spawn(async move {
+                        while let Ok(signal) = receiver.write().await.recv().await {
+                            let port_core_cloned = port_core.clone();
+                            let mut w_port_core = port_core_cloned.write().await;
+                            w_port_core.signal_hits += 1;
+                            Self::send_port_status(
+                                monitoring_store_cloned.clone(),
+                                &port_id,
+                                w_port_core.signal_hits,
+                                signal,
+                            )
+                            .await;
+                        }
+                    });
+
                     Ok(())
                 }
             }
-            // if  self.get
-            // let c_receiver = neuron.provide_output().await;
-            // let c_results = self.trace_log.clone();
-            // let c_output_interface = self.output_interface.clone();
-            // let c_port = network_port.clone();
-            // let jh = tokio::task::spawn(async move {
-            //     let port = Arc::new(c_port);
-            //     let mut w_port_result_receiver = c_receiver.write().await;
-            //     while let Ok(signal) = w_port_result_receiver.recv().await {
-            //         let output_width = c_output_interface.read().await.len();
-            //         let mut result: Vec<u8> = vec![0; output_width];
-            //         result[*port] = signal;
-            //         let result: String = result.iter().fold(String::new(), |mut acc, item| {
-            //             acc.push_str(item.to_string().as_str());
-            //             acc
-            //         });
-            //         c_results.write().await.push(result);
-            //     }
-            // });
-            // self.processing_registers
-            //     .write()
-            //     .await
-            //     .insert(network_port.clone(), jh);
         } else {
-            Err(Box::new(RnnError::IdNotFound))
+            Err(Box::new(RnnError::NeuronNotFound(neuron_id.to_string())))
         }
     }
 
     pub async fn get_output_receiver(&self, port: usize) -> Option<Arc<RwLock<Receiver<u8>>>> {
-        self.output_interface
-            .read()
-            .await
-            .get(&port)
-            .map(|rx| rx.clone())
+        if let Some(port_core) = self.output_interface.read().await.get(&port) {
+            let r_port_core = port_core.read().await;
+            if let SignalHandler::Output(receiver) = &r_port_core.signal_handler {
+                Some(receiver.clone())
+            } else {
+                None
+            }
+        } else {
+            None
+        }
     }
 
     pub fn get_id(&self) -> String {
         self.id.clone()
+    }
+
+    async fn send_port_status(
+        monitoring_store: Arc<RwLock<Vec<Status>>>,
+        port_id: &str,
+        signal_hits: u64,
+        recent_signal: Signal,
+    ) {
+        let mut w_monitoring_store = monitoring_store.write().await;
+        let timestamp = Utc::now();
+        w_monitoring_store.push(Status::Port(PortInfo {
+            timestamp,
+            id: port_id.to_string(),
+            hit_count: signal_hits,
+            recent_signal,
+        }));
     }
 }
 
@@ -485,7 +553,7 @@ mod tests {
         let neuron = net.create_neuron(net.clone(), 1, vec![]).await.unwrap();
         let id = neuron.get_id();
 
-        assert!(net.get_current_neuron_statistics(&id).await.is_ok());
+        assert!(net.get_current_neuron_status(&id).await.is_ok());
     }
 
     #[tokio::test]
@@ -499,7 +567,7 @@ mod tests {
         let res = net.connect_neurons(&src_id, &dst_id, 0).await;
         assert!(res.is_ok());
 
-        let src_stat = net.get_current_neuron_statistics(&src_id).await;
+        let src_stat = net.get_current_neuron_status(&src_id).await;
         assert!(src_stat.is_ok());
         let src_stat = match src_stat.unwrap() {
             Status::Neuron(info) => Some(info),
@@ -510,7 +578,7 @@ mod tests {
         assert_eq!(src_stat.dendrite_count, 1);
         assert_eq!(src_stat.dendrite_connected_count, 0);
 
-        let dst_stat = net.get_current_neuron_statistics(&dst_id).await;
+        let dst_stat = net.get_current_neuron_status(&dst_id).await;
         assert!(dst_stat.is_ok());
         let dst_stat = match dst_stat.unwrap() {
             Status::Neuron(info) => Some(info),
@@ -600,6 +668,24 @@ mod tests {
 
         tokio::time::sleep(Duration::from_millis(1)).await;
 
-        assert_eq!(net.pop_monitoring_store().await.len(), 1);
+        let monitoring_log = net.pop_monitoring_store().await;
+
+        assert_eq!(monitoring_log.len(), 3);
+        for status in monitoring_log {
+            match status {
+                Status::Neuron(neuron_info) => {
+                    assert!(neuron_info.timestamp.timestamp().is_positive());
+                    assert_eq!(neuron_info.accumulator, 1);
+                    assert_eq!(neuron_info.dendrite_connected_count, 1);
+                    assert_eq!(neuron_info.dendrite_count, 1);
+                    assert_eq!(neuron_info.dendrite_hit_count, 0);
+                    assert_eq!(neuron_info.receiver_count, 1);
+                }
+                Status::Port(port_info) => {
+                    assert!(port_info.timestamp.timestamp().is_positive());
+                    assert_eq!(port_info.hit_count, 1);
+                }
+            }
+        }
     }
 }